requirements.txt
# ── Core data & ML ──────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
imbalanced-learn==0.12.3
xgboost==2.0.3
lightgbm==4.3.0
shap==0.45.0
optuna==3.6.1
seaborn==0.13.2
# ── Deep learning (NLP + LSTM) ───────────────────────────────────
torch==2.3.0
transformers==4.41.0
datasets==2.19.1
accelerate==0.30.0
sentencepiece==0.2.0

# ── Time series ──────────────────────────────────────────────────
prophet==1.1.5

# ── API ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0

# ── Database & Cache ─────────────────────────────────────────────
sqlalchemy==2.0.30
pyodbc==5.1.0
aioodbc==0.5.0
alembic==1.13.1
redis==5.0.4
fastapi-cache2==0.2.1

# ── Dashboard ────────────────────────────────────────────────────
streamlit==1.35.0
plotly==5.22.0
streamlit-plotly-events==0.0.6

# ── MLOps ────────────────────────────────────────────────────────
mlflow==2.13.0

# ── External data ────────────────────────────────────────────────
yfinance==0.2.40
fredapi==0.5.2

# ── Dev & testing ────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
python-dotenv==1.0.1
notebook==7.2.0
ipykernel==6.29.4
black==24.4.2
ruff==0.4.4
